refactor(zk-explorer): Re-organize project structure, remove unused components
