+++ apps/zk-explorer/app/dashboard/components/DashboardOverviewTab.tsx
@@ -4,13 +4,13 @@ import { GithubFill, GlobalLine, TwitterFill } from '@webb-tools/icons';
 import { Button, Card, Typography } from '@webb-tools/webb-ui-components';
 import { WEBB_DOCS_URL } from '@webb-tools/webb-ui-components/constants';
 import { FC, useMemo } from 'react';
-import { useRequireAuth } from '../hooks/useAuth';
+import { LargeSquareAvatar } from '../../../components/LargeSquareAvatar';
+import { SocialChip } from '../../../components/SocialChip';
+import { useRequireAuth } from '../../../hooks/useAuth';
 import useTailwindBreakpoint, {
   TailwindBreakpoint,
-} from '../hooks/useTailwindBreakpoint';
-import { formatTimestamp } from '../utils';
-import { LargeSquareAvatar } from './LargeSquareAvatar';
-import { SocialChip } from './SocialChip';
+} from '../../../hooks/useTailwindBreakpoint';
+import { formatTimestamp } from '../../../utils';
 
 export const DashboardOverviewTab: FC = () => {
   const {
