--- apps/zk-explorer/components/CircuitCard/types.ts
@@ -1,7 +1,0 @@
-export type CircuitItem = {
-  ownerAvatarUrl: string;
-  filename: string;
-  description: string;
-  stargazerCount: number;
-  constraintCount: number;
-};
