+++ apps/zk-explorer/components/Filters/index.tsx
@@ -5,10 +5,10 @@ import assert from 'assert';
 import { cloneDeep } from 'lodash';
 import { FC, useCallback, useEffect, useState } from 'react';
 import { twMerge } from 'tailwind-merge';
+import { fetchFilterOptions } from '../../api/filters';
 import { IS_DEBUG_MODE } from '../../constants';
 import { MOCK_CATEGORIES } from '../../constants/mock';
 import { useFilterConstraints } from '../../hooks/useFilterConstraints';
-import { fetchFilterOptions } from '../../utils/api';
 import { FilterCheckboxItem } from '../FilterCheckboxItem';
 import { FilterCategoryItem, FilterConstraints } from './types';
 
@@ -18,7 +18,7 @@ export type FiltersProps = PropsOf<'div'> & {
   onConstraintsChange: (constraints: FilterConstraints) => void;
 };
 
-export const Filters: FC<FiltersProps> = ({
+const Filters: FC<FiltersProps> = ({
   className,
   hasCloseButton,
   onConstraintsChange,
@@ -120,3 +120,5 @@ export const Filters: FC<FiltersProps> = ({
     </div>
   );
 };
+
+export default Filters;
