--- apps/zk-explorer/components/GitHubAvatar/index.ts
@@ -1,1 +1,0 @@
-export { default as GitHubAvatar } from './GitHubAvatar';
