--- apps/zk-explorer/components/GitHubIconWithLink/index.ts
@@ -1,3 +1,0 @@
-import GitHubIconWithLink from './GitHubIconWithLink';
-
-export default GitHubIconWithLink;
