--- apps/zk-explorer/components/GitHubIconWithLink/types.ts
@@ -1,5 +1,0 @@
-import type { IconBase } from '@webb-tools/icons/types';
-
-export interface GitHubIconWithLinkProps extends IconBase {
-  href: string;
-}
