+++ apps/zk-explorer/components/SubmitPageControls.tsx
@@ -4,10 +4,10 @@ import { Button, Input } from '@webb-tools/webb-ui-components';
 import assert from 'assert';
 import { useRouter } from 'next/navigation';
 import { FC, useCallback, useState } from 'react';
+import { submitProject } from '../api/projects';
 import {
   createProjectDetailPath,
   parseGithubUrl,
-  submitProject,
   validateGithubUrl,
 } from '../utils';
