+++ apps/zk-explorer/containers/RelatedProjectsContainer/client/RelatedProjectsCarousel.tsx
@@ -8,7 +8,7 @@ import { ComponentProps, useMemo, type FC } from 'react';
 import { Pagination } from 'swiper/modules';
 import { Swiper, SwiperSlide } from 'swiper/react';
 
-import { ProjectCard } from '../../../components/ProjectCard/ProjectCard';
+import ProjectCard from '../../../components/ProjectCard';
 import type { ProjectItem } from '../../../components/ProjectCard/types';
 import { createProjectDetailPath } from '../../../utils';
