+++ apps/zk-explorer/context/SidebarContext.ts
