+++ apps/zk-explorer/utils/api.ts
@@ -1,14 +1,6 @@
 import { notificationApi } from '@webb-tools/webb-ui-components';
 import assert from 'assert';
-import { CircuitItem } from '../components/CircuitCard/types';
-import type {
-  FilterCategoryItem,
-  FilterConstraints,
-} from '../components/Filters/types';
-import { ProjectItem } from '../components/ProjectCard/types';
-import { Plan } from '../components/ProofGenerationStepCards/types';
 import { API_PREFIX } from '../constants';
-import { User } from '../hooks/useAuth';
 import { gracefullyParseJson } from './index';
 
 export enum ApiRoute {
@@ -32,20 +24,6 @@ export type ApiResponse<T = Record<string, never>> = {
   isSuccess: boolean;
   errorMessage?: string;
   data?: T;
-};
-
-export type ProjectSearchResponseData = {
-  projects: ProjectItem[];
-  resultCount: number;
-};
-
-export type CircuitSearchResponseData = {
-  circuits: CircuitItem[];
-  resultCount: number;
-};
-
-export type FilterOptionsResponseData = {
-  categories: FilterCategoryItem[];
 };
 
 export function makeAbsoluteApiEndpointUrl(route: ApiRoute): string {
@@ -130,156 +108,6 @@ export async function sendApiRequest<T = Record<string, never>>(
     innerResponse,
     fetchResponse,
   };
-}
-
-export async function exchangeAuthCodeForOAuthToken(
-  code: string
-): Promise<boolean> {
-  const responseWrapper = await sendApiRequest(
-    ApiRoute.OAuthGithub,
-    'GitHub auth code could not be exchanged for an OAuth token',
-    {
-      method: 'POST',
-      body: JSON.stringify({ code }),
-    }
-  );
-
-  // No need to handle the response further, as the OAuth
-  // token is set as an HTTP-only cookie by the backend.
-  return responseWrapper.innerResponse.isSuccess;
-}
-
-export enum SearchSortByClause {
-  MostPopular = 'Most Popular',
-  Newest = 'Newest',
-}
-
-export async function searchProjects(
-  constraints: FilterConstraints,
-  query: string,
-  page: number,
-  sortBy?: SearchSortByClause
-): Promise<ProjectSearchResponseData> {
-  const responseWrapper = await sendApiRequest<ProjectSearchResponseData>(
-    ApiRoute.SearchProjects,
-    'Failed to fetch projects',
-    {
-      method: 'GET',
-      body: JSON.stringify({
-        constraints,
-        query,
-        page,
-        sortBy: sortBy ?? null,
-      }),
-    }
-  );
-
-  return extractResponseData(responseWrapper);
-}
-
-export async function searchCircuits(
-  constraints: FilterConstraints,
-  query: string,
-  page: number,
-  sortBy?: SearchSortByClause
-): Promise<CircuitSearchResponseData> {
-  const responseWrapper = await sendApiRequest<CircuitSearchResponseData>(
-    ApiRoute.SearchCircuits,
-    'Failed to fetch circuits',
-    {
-      method: 'GET',
-      body: JSON.stringify({
-        constraints,
-        query,
-        page,
-        sortBy: sortBy ?? null,
-      }),
-    }
-  );
-
-  return extractResponseData(responseWrapper);
-}
-
-export async function submitProject(githubSlug: string): Promise<ApiResponse> {
-  const responseWrapper = await sendApiRequest(
-    ApiRoute.Projects,
-    'Failed to submit project',
-    {
-      method: 'POST',
-      body: JSON.stringify({ githubSlug }),
-    }
-  );
-
-  return responseWrapper.innerResponse;
-}
-
-export async function fetchFilterOptions(): Promise<FilterOptionsResponseData> {
-  const responseWrapper = await sendApiRequest<FilterOptionsResponseData>(
-    ApiRoute.Constraints,
-    'Failed to fetch filter options',
-    {
-      method: 'GET',
-    }
-  );
-
-  // TODO: Temporary; Using `assert` here is incorrect, as this would not necessarily equate to a logic error.
-  assert(
-    responseWrapper.innerResponse.data !== undefined,
-    'Response data should not be undefined'
-  );
-
-  return responseWrapper.innerResponse.data;
-}
-
-export async function updateUserProfile(
-  changes: Partial<User>
-): Promise<ApiResponse> {
-  const responseWrapper = await sendApiRequest(
-    ApiRoute.User,
-    'Failed to update user profile settings',
-    {
-      method: 'PUT',
-      body: JSON.stringify(changes),
-    }
-  );
-
-  return responseWrapper.innerResponse;
-}
-
-export async function requestProofGeneration(data: {
-  r1cs: File;
-  plan: Plan;
-  verificationKey: File;
-  provingKey: File;
-  mpcParticipantAddresses: string[];
-}): Promise<ApiResponse> {
-  const formData = new FormData();
-
-  // TODO: Need to centralize these keys somewhere more accessible, so that they can easily be found and updated later on.
-  formData.append('r1cs', data.r1cs);
-  formData.append('plan', JSON.stringify(data.plan));
-  formData.append('verificationKey', data.verificationKey);
-  formData.append('provingKey', data.provingKey);
-
-  formData.append(
-    'mpcParticipantAddresses',
-    JSON.stringify(data.mpcParticipantAddresses)
-  );
-
-  const responseWrapper = await sendApiRequest(
-    ApiRoute.ProofGeneration,
-    'Failed to request proof generation',
-    {
-      method: 'POST',
-      body: formData,
-      // Do not specify the 'Content-Type' header, as it
-      // will be automatically set by the browser. The browser
-      // will automatically add a 'boundary' parameter to the header,
-      // which is required for multipart form data.
-    }
-  );
-
-  return responseWrapper.innerResponse;
 }
 
 export function extractResponseData<T>(
